enhanced_requirements.txt
# Enhanced Royalton AI Screener Requirements
# Install with: pip install -r enhanced_requirements.txt

# Core requirements (already in main requirements)
pandas>=1.5.0
openpyxl>=3.0.0
pdfminer.six>=20220524
python-docx>=0.8.11

# Enhanced OCR capabilities
pytesseract>=0.3.10
Pillow>=9.0.0
pdf2image>=3.1.0

# Semantic matching with spaCy
spacy>=3.4.0

# Configuration management
PyYAML>=6.0

# Web interface
streamlit>=1.25.0

# Optional: For better document parsing
python-magic>=0.4.27
python-magic-bin>=0.4.14; platform_system=="Windows"

# Optional: For advanced NLP features
sentence-transformers>=2.2.0
scikit-learn>=1.1.0
